requirements.txt
requests
termcolor
discord.py
rethinkdb
jishaku
lavalink